fix compile problem in clang 8

+++ atframework/service/component/etcdcli/etcd_cluster.cpp
@@ -127,7 +127,8 @@ namespace atframe {
             }
 
             EXPLICIT_UNUSED_ATTR static int etcd_cluster_verbose_callback(util::network::http_request &req, curl_infotype type, char *data, size_t size) {
-                if (util::log::log_wrapper::check_level(WDTLOGGETCAT(util::log::log_wrapper::categorize_t::DEFAULT), util::log::log_wrapper::level_t::LOG_LW_TRACE)) {
+                if (util::log::log_wrapper::check_level(WDTLOGGETCAT(util::log::log_wrapper::categorize_t::DEFAULT),
+                                                        util::log::log_wrapper::level_t::LOG_LW_TRACE)) {
                     const char *verbose_type = "Unknown Action";
                     switch (type) {
                     case CURLINFO_TEXT:
@@ -783,7 +784,8 @@ namespace atframe {
                 rapidjson::Document doc;
                 doc.SetObject();
                 doc.AddMember("ID", get_lease(), doc.GetAllocator());
-                doc.AddMember("TTL", std::chrono::duration_cast<std::chrono::seconds>(conf_.keepalive_timeout).count(), doc.GetAllocator());
+                doc.AddMember("TTL", static_cast<int64_t>(std::chrono::duration_cast<std::chrono::seconds>(conf_.keepalive_timeout).count()),
+                              doc.GetAllocator());
 
                 setup_http_request(req, doc, get_http_timeout_ms(), conf_.authorization_header);
                 req->set_priv_data(this);
@@ -1177,7 +1179,8 @@ namespace atframe {
 
             // req->set_on_verbose(details::etcd_cluster_verbose_callback);
 
-            if (util::log::log_wrapper::check_level(WDTLOGGETCAT(util::log::log_wrapper::categorize_t::DEFAULT), util::log::log_wrapper::level_t::LOG_LW_TRACE)) {
+            if (util::log::log_wrapper::check_level(WDTLOGGETCAT(util::log::log_wrapper::categorize_t::DEFAULT),
+                                                    util::log::log_wrapper::level_t::LOG_LW_TRACE)) {
                 req->set_on_progress(details::etcd_cluster_trace_porcess_callback);
             }
